refactor: move provider and job files into src/providers/ and src/jobs/
Moved into src/providers/: embeddings.ts, reranker.ts, llm-provider.ts,
vector-store.ts, expansion.ts. Moved into src/jobs/: watcher.ts,
consolidation.ts, consolidation-worker.ts.

Each original path replaced with 1-line barrel shim (export *) so all
existing import paths continue to work unchanged.

Updates test/rri-t-round4.test.ts to point to new provider/job paths.
Also adds OpenSpec proposal/design/specs/tasks for Phase 4.

Zero behavior changes. Tests: 1518 passing (10 pre-existing watcher failures).
