+++ openspec/changes/nano-brain-phase4-providers-jobs/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-05
